Disable CFI in __apply_alternatives function.
Normally, when the kernel takes the address of a function in order to call
it, the address is taken by C, which allows the compiler to insert a jump
table entry for the function. Here is one exception. The __apply_alternatives
function indirectly calls a function whose address is taken by assembly
(specifically, arm64_enable_wa2_handling, whose address is taken in
arch/arm64/kernel/entry.S). With my prototype compiler changes, this CFI
check will fail because no jump table entry is created.

For now, disable CFI in this function.