arm64: mte: optimize asynchronous tag check fault flag check
We don't need to check for MTE support before checking the flag
because it can only be set if the hardware supports MTE. As a result
we unconditionally check the flag bit which is expected to be in a
register and therefore the check can be done in a single instruction
instead of first needing to load the hwcaps.

On a DragonBoard 845c with a kernel built with CONFIG_ARM64_MTE=y with
the powersave governor this reduces the cost of a kernel entry/exit
(invalid syscall) from 465.1ns to 463.8ns.